% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclass.R
\docType{class}
\name{BioMDA-class}
\alias{BioMDA-class}
\title{BioMDA class}
\description{
BioMDA class
}
\section{Slots}{

\describe{
\item{\code{metabolite}}{The synonyms of compound.}

\item{\code{metaboliteInfo}}{A list containing biological prior knowledge of query
metabolite.}

\item{\code{disease}}{The synonyms of disease.}

\item{\code{diseaseInfo}}{A list containing biological prior knowledge of query
disease.}

\item{\code{branch}}{The branch of BioMDA pipeline, one of "tgtp", "ssimm" and
"coabm".}

\item{\code{score}}{Threshold of significance to include a interaction, a number
between 0 and 1000 (Highest >= 900, High >= 700, medium >= 400 and low >=
150. Default: 700).}

\item{\code{node}}{Network nodes represent proteins or metabolites.}

\item{\code{edge}}{Edges represent protein-protein, compound-protein, or
compound-compoud associations.}

\item{\code{diseaseGeneORA}}{A list containing the result of disease-related genes
ORA.}

\item{\code{enrichmentResult}}{A list of \code{enrichResult} object.
imported from DOSE package.}
}}

