% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomda-package.R
\docType{package}
\name{biomda-package}
\alias{biomda}
\alias{biomda-package}
\title{biomda: A computational Framework for predictive Metabolite-Disease Association mapping through Biomedical knowledge integration}
\description{
The biomda package provides a systematic, data-driven approach to infer potential metabolite-disease associations by leveraging heterogeneous biomedical knowledge sources. This tool integrates multi-omics datasets, curated pathway databases, and literature-mined evidence to construct a probabilistic network for hypothesis generation.
}
\author{
\strong{Maintainer}: Li Zhan \email{smu18575877413@gmail.com} (\href{https://orcid.org/0009-0003-7470-7586}{ORCID})

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [contributor]
  \item Huimin Zheng \email{zhenghuimin91@126.com} (\href{https://orcid.org/0000-0003-3489-0964}{ORCID}) [contributor]
}

}
\keyword{internal}
