% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclass.R
\name{BioMDA}
\alias{BioMDA}
\title{BioMDA object}
\usage{
BioMDA(metabolite, tgtp_score = 700, coab_metabolites = NULL, disease = NULL)
}
\arguments{
\item{metabolite}{The synonyms of metabolite/compound (ignored case), e.g.
"aspirin" or some external id.}

\item{tgtp_score}{Threshold of significance to search known metabilite-
protein interactions (i.e., target proteins of query metabilite), a number
between 0 and 1000 (Highest >= 900, High >= 700, medium >= 400 and low >=
150. Default: 700).}

\item{coab_metabolites}{A list of co-abundant metabolites of query metabolite
(i.e., consistency abundance module), identified by WGCNA.}

\item{disease}{The synonyms of disease, e.g. "Crohn' disease" or "DOID:8778"
(Disease Ontology ID). When the exact disease name is uncertain, it may be
identified using \code{disease2doid("crohn", fixed = FASLE)} for fuzzy
matching.}
}
\value{
BioMDA object.
}
\description{
Construct a BioMDA object.
}
\details{
Some external id of "aspirin":
\itemize{
\item \code{KEGG compound} D00109
\item \code{ChEBI} CHEBI:15365
\item \code{HMDB} HMDB0001879
\item \code{CAS-RN} 50-78-2
\item etc.
}

The descriptions of metaboliteInfo:
\itemize{
\item \code{$cid} the PubChem Cid of metabolite.
\item \code{$target_proteins} the target proteins of metabolite, accessed by
\code{tgtp_search}.
\item \code{$ssim_metabolites} a set of metabolites shared similar structure with
query metabolite, accessed by \code{ssimcid_search}.
\item \code{$coab_metabolites} a set of co-abundant metabolites of query
metabolite (i.e., consistency abundance module), identified by WGCNA.
}

The descriptions of diseaseInfo:
\itemize{
\item \code{$doid} Disease Ontology ID of disease.
\item \code{superclass} superclass of disease.
\item \code{disease_genes} a data.frame of disease-related genes, accessed by
\code{disgene_search}.
}

The descriptions of diseaseGeneORA:
\itemize{
\item \code{$disease} Synonyms of disease.
\item \code{$geneRatio} (Number of interaction genes in the disease-related
genes) / (Total number of interaction genes)
\item \code{$bgRatio} (Number of disease-related genes)
/ (Total number of background genes, i.e., universe)
\item \code{$pvalue} P value of hypergeometric test.
\item \code{$geneID} The Overlap genes between genes and disease-related genes.
\item \code{$count} Number of interaction genes in the disease-related genes.
}

The descriptions of enrichmentResult, a list containing:
\itemize{
\item \code{$`Annotated Keywords (UniProt)`}
\item \code{$`Biological Process (Gene Ontology)`}
\item \code{$`Cellular Component (Gene Ontology)`}
\item \code{$`Molecular Function (Gene Ontology)`}
\item \code{$`Disease-gene associations (DISEASES)`}
\item \code{$`Human Phenotype (Monarch)`}
\item \code{$`KEGG Pathways`}
\item \code{$`Local Network Cluster (STRING)`}
\item \code{$`Protein Domains (Pfam)`}
\item \code{$`Protein Domains (SMART)`}
\item \code{$`Protein Domains and Features (InterPro)`}
\item \code{$`Reactome Pathways`}
\item \code{$`Subcellular localization (COMPARTMENTS)`}
\item \code{$`Tissue expression (TISSUES)`}
\item \code{$`WikiPathways`}
\item \code{$`Disease Gene Network (DisGeNET)`}
\item \code{$`Network of Cancer Genes (NCG)`}
\item \code{$`Disease Ontology (DO)`}
}
}
\examples{
library(biomda)
# Query for diseases containing "crohn".
disease2doid("crohn", fixed = FALSE)
##   doid      disease         superclass                 gene_count
##   <chr>     <chr>           <chr>                           <int>
## 1 DOID:8778 Crohn's disease inflammatory bowel disease        110
biomda <- BioMDA(metabolite = "Urobilin", disease = "Crohn's disease")
}
\seealso{
\code{BioMDA-class}
}
