% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-biomda-branch.R
\name{biomda_branch}
\alias{biomda_branch}
\alias{biomda_branch,BioMDA-method}
\alias{biomda_branch,BioMDA}
\title{BioMDA branch}
\usage{
biomda_branch(obj, branch = c("tgtp", "ssimm", "coabm"), score = 700)

\S4method{biomda_branch}{BioMDA}(obj, branch = c("tgtp", "ssimm", "coabm"), score = 700L)
}
\arguments{
\item{obj}{BioMDA object}

\item{branch}{The branch name, one of "tgtp", "ssimm" and "coabm".}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (Highest >= 900, High >= 700, medium >= 400 and low >=
150. Default: 700).}
}
\value{
BioMDA object containing result of branch.
}
\description{
The pipeline of BioMDA to predict potential metabolite-disease
associations in three different branches. Ref: fig 1.C in Huimin Zheng, et
al. (2022).
}
\details{
The branch of BioMDA pipeline:
\itemize{
\item \code{tgtp} the target proteins branch.
\item \code{ssimm} the structurally similar metabolites branch.
\item \code{coabm} the co-abundant metabolites branch.
}
}
\examples{
library(biomda)
biomda <- BioMDA("Urobilin", disease = "Crohn's disease")
ssimm_biomda <- biomda_branch(biomda, branch = "ssimm", score = 700)
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
