% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utilities.R
\name{biomda_db_api}
\alias{biomda_db_api}
\title{APIs of BioMDA Database}
\usage{
biomda_db_api(resource, .params = list())
}
\arguments{
\item{resource}{The resource locator, e.g. "/c2cid/".}

\item{.params}{A list includes the name-value parameters will process the
form.}
}
\value{
The text from the HTTP response or NULL.
}
\description{
Submit an HTML form to APIs of BioMDA Database using the POST
method.
}
\examples{
library(biomda)
protein <- c("PTCH1", "TP53", "BRCA1", "BRCA2")
resp <- biomda_db_api(resource = "/protein2stringid/",
                      .params = list(protein = protein))
resp
## $p2string
##   protein       stringv12
## 1    TP53 ENSP00000269305
## 2   PTCH1 ENSP00000332353
## 3   BRCA2 ENSP00000369497
## 4   BRCA1 ENSP00000418960
}
