% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-disease-gene-ora.R
\name{biomda_disge_ora}
\alias{biomda_disge_ora}
\alias{biomda_disge_ora,BioMDA-method}
\alias{biomda_disge_ora,BioMDA}
\title{Disease-related genes ORA}
\usage{
biomda_disge_ora(obj, universe = 21306)

\S4method{biomda_disge_ora}{BioMDA}(obj, universe = 21306)
}
\arguments{
\item{obj}{BioMDA object containing result of branch, see
\code{\link{biomda_branch}}.}

\item{universe}{Count of universal background protein-coding genes.
For example, background of homo sapiens is 21306.}
}
\value{
BioMDA object containing result of disease-related ORA.
}
\description{
Determine whether genes from disease-related genes are enriched
more than would be expected in potential interaction genes of metabolite.
Ref: fig 1.C in Huimin Zheng, et al. (2022).
}
\details{
A list contained:
\itemize{
\item \code{disease} Synonyms of disease.
\item \code{geneRatio} (Number of interaction genes in the disease-related genes)
/ (Total number of interaction genes)
\item \code{bgRatio} (Number of disease-related genes)
/ (Total number of background genes, i.e., universe)
\item \code{pvalue} P value of hypergeometric test.
\item \code{geneID} The Overlap genes between genes and disease-related genes.
\item \code{count} Number of interaction genes in the disease-related genes.
}
}
\examples{
library(biomda)
biomda <- BioMDA("Urobilin", "Crohn's disease")
ssimm_biomda <- biomda_branch(biomda, branch = "ssimm", score = 700)
ssimm_biomda <- biomda_disge_ora(ssimm_biomda)
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
