% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-id-translator.R
\name{compound2cid}
\alias{compound2cid}
\title{Convert synonyms of compound to PubChem cid.}
\usage{
compound2cid(compound)
}
\arguments{
\item{compound}{Synonyms of compound (ignored case), e.g. "aspirin" or
some external id (see details).}
}
\value{
A data.frame containing queried compound and PubChem cid.
}
\description{
Convert synonyms of compound to PubChem cid.
}
\details{
some external id of "aspirin":
\itemize{
\item \code{KEGG compound} D00109
\item \code{ChEBI} CHEBI:15365
\item \code{HMDB} HMDB0001879
\item \code{CAS-RN} 50-78-2
\item etc.
}
}
\examples{
library(biomda)
compound <- c("N-methylserotonin", "Urobilin", "Ononetin")
c2cid <- compound2cid(compound = compound)
}
