% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-abun-analysis.R
\name{diff_metabolites}
\alias{diff_metabolites}
\title{Differential abundance analysis}
\usage{
diff_metabolites(
  abundance,
  resp,
  compared,
  ortho = FALSE,
  log10L = FALSE,
  scaling = "standard",
  test = "t.test",
  padj_method = "BH",
  vip_cutoff = 1,
  padj_cutoff = 0.05,
  auc_cutoff = 0.7,
  log2fc_cutoff = 1.5,
  ...
)
}
\arguments{
\item{abundance}{Abundance table, a matrix or data frame in which columns are
metabolites and rows ar samples.}

\item{resp}{Response to be modelled, a factor (same length as row number of
\code{abundance}) with two levels.}

\item{compared}{Define the condition A and condition B in "response" for
calculating fold change. For example, set "compared" to c("case", "control"),
then the fold change is "case/control".}

\item{ortho}{FALSE (for PLS-DA) or TRUE (for OPLS-DA).}

\item{log10L}{Whether the 'Abudnace Table' table needs to be
log10-transformed.}

\item{scaling}{Scaling methods for "abundance", one of "none", "center",
"pareto", and "standard" (default). See details.}

\item{test}{t.test or wilcox.test.}

\item{padj_method}{P-values adjusted methods, one of "holm", "hochberg",
"hommel", "bonferroni", "BH", "BY", "fdr" and "none".}

\item{vip_cutoff}{Threshold of VIP score of the (O)PLS-DA model.}

\item{padj_cutoff}{Threshold of P-value adjusted for t.test (or wilcox.test).}

\item{auc_cutoff}{Threshold of auc of ROC curve.}

\item{log2fc_cutoff}{Threshold of absolute value of log2 fold change.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A data.frame with fields:
\itemize{
\item \code{metabolite} Metabolites
\item \code{vip} Variable Importance in Projection (VIP) of (O)PLS-DA module
\item \code{auc} AUC of ROC curve analysis
\item \code{pval, padj} P value and p value adjusted of t.test (or wilcox.test)
\item \code{fc, log2fc, fc_direct} Fold change (FC), log2(FC) and direction of
FC (e.g., case/control).
\item \code{significant} 1 for significantly differential metabolites, or 0 for
not significant.
}
}
\description{
Differential abundance analysis for metabolomics.
}
\details{
Scaling methods as following:
\itemize{
\item \code{none} none
\item \code{center}  mean-centering scaling
\item \code{pareto} mean-centering and pareto scaling
\item \code{standard} mean-centering and unit variance scaling, default
}
}
\note{
Fold change was calculated using the \code{original abundance} table,
while p-values and AUC values were computed based on log10-transformed (when
\code{log10L = TRUE}, or FALSE for untransformed) abundance data. VIP scores
were derived from scaled data (default: \code{scaling = "standard"}, see
details) using either log10-transformed or untransformed abundance values.
}
\examples{
library(biomda)
data(IBD.metabolomics)
## Discovery cohort (PRISM)
ibd.discovery <- IBD.metabolomics$discovery
abunda <- ibd.discovery$abundance
metada <- ibd.discovery$metadata
# CD v.s. Control
metada <- metada[metada$Diagnosis != "UC", ]
table(metada$Diagnosis) # 68 Crohn' disease and 34 control samples
abunda <- abunda[match(rownames(metada), rownames(abunda)), ]
# OPLS-DA
compared <- c("CD", "Control")
diffres <- diff_metabolites(abunda, metada$Diagnosis, compared, ortho = TRUE)
}
