% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-id-translator.R
\name{disease2doid}
\alias{disease2doid}
\title{Convert aliases of disease to Diease Ontology id (DOID).}
\usage{
disease2doid(disease, fixed = TRUE)
}
\arguments{
\item{disease}{Aliases of disease(ignored case), e.g. "pre-eclampsia". Note:
the length of "disease" must be 1.}

\item{fixed}{If TRUE match "disease" exactly, otherwise query "disease" fuzzy
faintly.}
}
\value{
A data.frame containing DOID, name, superclass and number of related
genes of the queried disease.
}
\description{
Convert aliases of disease to Diease Ontology id (DOID).
}
\examples{
library(biomda)
# Query for pre-eclampsia disease (PE).
disease2doid("pre-eclampsia")
# Query by directly entering the DOID.
disease2doid("DOID:10591")
# Query for diseases containing "obesity".
disease2doid("obesity", fixed = FALSE)
}
