% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-disease-genes.R
\name{disgene_search}
\alias{disgene_search}
\title{Retrieve disease-related genes.}
\usage{
disgene_search(doid)
}
\arguments{
\item{doid}{Disease Ontology ID (DOID). Note: the length of "doid" must be 1.}
}
\value{
A tibble with ncbi entrez gene id and gene symbol.
}
\description{
Retrieve disease-related genes.
}
\details{
The aliases of disease can be converted to DOID using
\code{\link{disease2doid}}.
}
\examples{
library(biomda)
# Query for pre-eclampsia disease (PE, DOID:10591).
pe_disge <- disgene_search("DOID:10591")
# Query for morbid obesity disease (DOID:11981).
morbid_obesity_disge <- disgene_search("DOID:11981")
}
