% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{IBD.metabolomics}
\alias{IBD.metabolomics}
\title{(Data) Untargeted LC–MS metabolomics for IBD cohort}
\description{
A 155-member discovery cohort and a 65-member validation cohort,
each containing a crosssectional sampling of UC, CD and control patients.
\itemize{
\item \code{discovery} Discovery cohort (PRISM)
\item \code{validation} Validation cohort (LifeLines DEEP and NLIBD)
Ref: Eric A. Franzosa, et al. (2019).
}
}
\examples{
data(IBD.metabolomics)
## Discovery cohort (PRISM)
discovery <- IBD.metabolomics$discovery
## Validation cohort (LifeLines DEEP and NLIBD)
validation <- IBD.metabolomics$validation
}
\references{
Eric A. Franzosa, et al. (2019). Gut microbiome structure and
metabolic activity in inflammatory bowel disease.
}
\keyword{data}
