% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-id-translator.R
\name{protein2stringid}
\alias{protein2stringid}
\title{Convert synonyms of protein to STRING id.}
\usage{
protein2stringid(protein)
}
\arguments{
\item{protein}{Synonyms of protein(ignored case), e.g. "TP53".}
}
\value{
A data.frame containing queried protein and STRING id.
}
\description{
Convert synonyms of protein to STRING id.
}
\examples{
library(biomda)
protein <- c("PTCH1", "TP53", "BRCA1", "BRCA2")
p2stringid <- protein2stringid(protein = protein)
}
