% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.R
\name{show,BioMDA-method}
\alias{show,BioMDA-method}
\title{Show method extensions to show for BioMDA object.}
\usage{
\S4method{show}{BioMDA}(object)
}
\arguments{
\item{object}{BioMDA object}
}
\value{
Print info of BioMDA object.
}
\description{
Show method extensions to show for BioMDA object.
}
