% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-stitch-network.R
\name{stitch_network}
\alias{stitch_network}
\title{STITCH network}
\usage{
stitch_network(cid, score = 700)
}
\arguments{
\item{cid}{PubChem cid, e.g. 'CID2244' ('aspirin').}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (Highest >= 900, High >= 700, medium >= 400 and low >=
150. Default: 700).}
}
\value{
A list containing weights of each edge and the annotations of each
node as following:

the tibble \code{edges} with fields:
\itemize{
\item \code{node1, node2} protein- (ENSPxxx) or compound-nodes (CIDxxx) in the
network.
\item \code{type} type of interaction edges in network, e.g., cpi
(compound-protein interactions), cci (compound-compound interactions) and ppi
(protein-protein interactions).
\item \code{score} combined score.
\item \code{experiment} experimentally determined (rank: known interactions).
\item \code{database} from curated databases (rank: known interactions).
\item \code{neighborhood} gene neighborhood (rank: predicted interactions).
\item \code{fusion} gene fusions (rank: predicted interactions).
\item \code{cooccurence} gene co-occurrence (rank: predicted interactions).
\item \code{structure} extrapolated 'compound-protein associations' based on
chemical structure (rank: predicted interactions).
\item \code{coexpression} co-expression (rank: others).
\item \code{textmining} textmining (rank: others).
}

the tibble \code{nodes} with fields:
\itemize{
\item \code{id} PubChem cid (compound) or STRING id (protein).
\item \code{type} type of nodes in network, protein or metabolite.
\item \code{name} common synonyms of compound or protein.
\item \code{external} external id, NA (for compounds) or Entrez Gene ID (for
proteins).
\item \code{description} description of compound or protein.
}
}
\description{
STITCH compound-protein, compound-compoud, interactions network.
}
\details{
The synonyms of compound can be converted to PubChem cid using
\code{\link{compound2cid}}.
}
\examples{
library(biomda)
# Use the structurally similar metabolites branch
cid <- compound2cid("ononetin") # "CID259632"
ssimcid <- ssimcid_search(cid)$ssimcid
stitchnw <- stitch_network(ssimcid, score = 700)
stitchnw$edges
stitchnw$nodes
}
