% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-string-network.R
\name{string_network}
\alias{string_network}
\title{STRING network}
\usage{
string_network(stringid, score = 700)
}
\arguments{
\item{stringid}{STRING id, e.g. ENSP00000269305.}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (Highest >= 900, High >= 700, medium >= 400 and low >=
150. Default: 700).}
}
\value{
A list containing weights of each edge and the annotations of each
node as following:

the tibble \code{edges} with fields:
\itemize{
\item \code{node1, node2} protein- (ENSPxxx) or compound-nodes (CIDxxx) in the
network.
\item \code{type} type of interaction edges in network, e.g., cpi
(compound-protein interactions), cci (compound-compound interactions) and ppi
(protein-protein interactions).
\item \code{score} combined score.
\item \code{experiment} experimentally determined (rank: known interactions).
\item \code{database} from curated databases (rank: known interactions).
\item \code{neighborhood} gene neighborhood (rank: predicted interactions).
\item \code{fusion} gene fusions (rank: predicted interactions).
\item \code{cooccurence} gene co-occurrence (rank: predicted interactions).
\item \code{structure} extrapolated 'compound-protein associations' based on
chemical structure (rank: predicted interactions).
\item \code{coexpression} co-expression (rank: others).
\item \code{textmining} textmining (rank: others).
}

the tibble \code{nodes} with fields:
\itemize{
\item \code{id} PubChem cid (compound) or STRING id (protein).
\item \code{type} type of nodes in network, protein or metabolite.
\item \code{name} common synonyms of compound or protein.
\item \code{external} external id, NA (for compounds) or Entrez Gene ID (for
proteins).
\item \code{description} description of compound or protein.
}
}
\description{
STRING proterin-protein interactions network.
}
\details{
The synonyms of protein can be converted to STRING id using
\code{\link{protein2stringid}}.
}
\examples{
library(biomda)
# Use the target proteins branch
# The ononetin and its catabolites inhibit TRPM3 and activate TRPA1.
stringid <- protein2stringid(c("TRPM3", "TRPA1"))$stringid
stringnw <- string_network(stringid, score = 700)
stringnw$edges
stringnw$nodes
}
