% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-target-proteins.R
\name{tgtp_search}
\alias{tgtp_search}
\title{Target proteins of metabolite}
\usage{
tgtp_search(cid, score = 700)
}
\arguments{
\item{cid}{PubChem cid, e.g. 'CID2244' ('aspirin').}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (Highest >= 900, High >= 700, medium >= 400 and low >=
150. Default: 700).}
}
\value{
A tibble containing the queried Compound (cid) along with
comprehensive information on target proteins, including STRING database IDs
(stringid), interaction scores (score), protein names (name), Entrez gene IDs
(entrezid), and detailed protein descriptions.
}
\description{
Search known metabolite-protein interactions (experimentally
determined or from curated databases) in STITCH.
}
\details{
The synonyms of compound can be converted to PubChem cid using
\code{\link{compound2cid}}.
}
\examples{
library(biomda)
cid <- compound2cid("N-methylserotonin")$cid # "CID150885"
cid2tgtp <- tgtp_search(cid)
}
